+++ examples/01_create_basic_thz.py
@@ -12,17 +12,14 @@ if __name__ == "__main__":
     with DotthzFile(path1, "w") as file:
         file.create_measurement("Measurement 1")
 
-        # create meta-data
-        meta_data = DotthzMetaData()
-        meta_data.user = "John Doe"
-        meta_data.version = "1.00"
-        meta_data.instrument = "Toptica TeraFlash Pro"
-        meta_data.mode = "THz-TDS/Transmission"
+        # create metadata
+        metadata = DotthzMetaData()
+        metadata.user = "John Doe"
+        metadata.version = "1.00"
+        metadata.instrument = "Toptica TeraFlash Pro"
+        metadata.mode = "THz-TDS/Transmission"
 
-        file["Measurement 1"].set_meta_data(meta_data)
+        file["Measurement 1"].set_metadata(metadata)
 
         # for thzVer 1.00, we need to transpose the array!
-
-        # important: do not manipulate keys on the `dataset` field, otherwise
-        # it won't be written to the file.
         file["Measurement 1"]["Sample"] = np.array([time, data]).T